requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
Flask-Limiter>=3.5.0
click>=8.1.0
rich>=13.7.0
browser-cookie3>=0.19.1
beautifulsoup4>=4.12.0
